Move the queue service into resources/js
QueueService and the queue types are Control Panel application code, not
reusable UI-library material, so they move from @craftcms/ui to
resources/js/common (services/Queue.ts, types/queue.ts). The CpServices
global interface moves with them since it types the app-bootstrapped
$queue/$config globals.

The moved service keeps relative/bare-package imports only, because the
legacy webpack bundle (CP.js) also compiles it and doesn't know Vite's @/
alias. Its test now runs under the root vitest suite.
