+++ resources/js/bootstrap/cp.ts
@@ -1,4 +1,5 @@
-import {QueueService, ConfigService} from '@craftcms/ui';
+import {ConfigService} from '@craftcms/ui';
+import {QueueService} from '@/common/services/Queue';
 import {createInertiaApp, router} from '@inertiajs/vue3';
 import QueueManager from '@/modules/utilities/components/queue-manager/QueueManager.vue';
 import {Axios, Config, Queue} from '@/common/types/keys';
