+++ resources/js/modules/utilities/components/queue-manager/RetryJobButton.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
   import {t} from '@craftcms/ui';
-  import {type JobInfo, JobStatus} from '@craftcms/ui';
+  import {type JobInfo, JobStatus} from '@/common/types/queue';
   import {useActionClient} from '@/common/composables/useFetch';
   import {unref, watch} from 'vue';
   import {router} from '@inertiajs/vue3';
